Smartpack resprite by iamtheguy (#8202) Attempt 3 of adding the iamtheguy smartpack reskins, now flavored as the S-V42A1 upgrade to the normal smartpack. (both of which are just cosmetic now.)
Previous prs: https://github.com/cmss13-devs/cmss13/pull/7207,
https://github.com/cmss13-devs/cmss13/pull/6669

The old smartpack is (in my opinion) still within standards so it still
exists, this exists as a further option.

The smartpack exists as a fun aesthetic item that allows synths to not
just creep around in the dark spooking marines, but it still a
mechanical downgrade over other backpacks, this brings in some new cool
looking sprites to choose from, since a lot of synth players are not
super fond of the current one.

<details>
<summary>Screenshots & Videos</summary>

![image](https://github.com/user-attachments/assets/8fa36e1d-63ff-404d-a412-f9795df492f2)

![image](https://github.com/user-attachments/assets/9d42e3b9-0587-44e0-aa71-49e99470b596)

![image](https://github.com/user-attachments/assets/201ae478-a012-4aa3-ae94-afd6a70793b1)

![image](https://github.com/user-attachments/assets/f04fd5bf-6f51-4207-9299-9ac6fc8d84b9)

</details>
:cl: iamtheguy, quicklode
imageadd: new iamtheguy smartpack sprites are now in the synth vendor.
/:cl: